Included input data files in .gitignore file.
There is one input data file per city included in the bikeshare study.
They are large .csv files, therefore they must be left out of commits to
local repo and pushes to the remote repo.  They were placed in the
folder input_data_files for neatness and the folder input_data_files was
listed in the .gitignore file.