Edited docustring for the function in the Bikeshare project file that filters the input csv files to correctly specify what happens if the user specifies "all" when prompted to filter by a dimension's values (e.g. cities).
This commit preceeds a the modification to Bikeshare_Prj_Huw.py's code
that I will make in the subsequent commit to allow the user to enter str
"all" instead of a comma seperated list of all of the values of a
dimension if they don't want to filter on specific values of that
dimension. Saves time for the user.