Edit code to correctly use input "all" when a user is prompted to provide the values they want to see for a dimension in the case where they don't want to filter on specific values (e.g. they wants stats for data points over all months in the bike system trial).
This improves user experience by not forcing the user to type out all
values for a dimension if they don't want to filter. Specifically edit
the function user_input() to correctly take "all" as a user input and
then the funnction that filters the input data load_data2() to correctly
recognise "all" as a parameter to then not filter the data.